pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "segvigen"
version = "0.1.0"
description = "3D mesh part segmentation using SegviGen flow-matching models"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = [
    "torch>=2.0",
    "numpy",
    "trimesh[easy]",
    "Pillow",
    "tqdm",
    "huggingface-hub",
]

[project.optional-dependencies]
# Install with: pip install segvigen[dev]
dev = ["uvicorn", "fastapi", "pydantic"]

[tool.setuptools.packages.find]
where = ["."]
include = ["segvigen*"]
